// apps/web/src/services/generated/models/businessRoleVo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API Documentation
 *
 *     当前环境: development
 *     最后更新: 9/15/2026, 5:29:26 PM
 *     [项目设计文档](https://wiki.example.com)
 *
 * OpenAPI spec version: 1.0
 */
import type { BusinessRoleParentVo } from "./businessRoleParentVo";

export interface BusinessRoleVo {
  /** 业务角色 ID */
  id: string;
  /** 业务角色标识 */
  key: string;
  /** 业务角色名称 */
  label: string;
  /** 角色说明 */
  description: string;
  /** 权限级别 */
  level: number;
  /**
   * 父角色 ID
   * @nullable
   */
  parentId?: string | null;
  /** 创建时间 */
  createdTime?: string;
  /** 更新时间 */
  updatedTime?: string;
  /**
   * 父角色
   * @nullable
   */
  parent?: BusinessRoleParentVo | null;
}